+++ src/commands/index.ts
@@ -9,6 +9,7 @@ import { autowork } from './autowork.js';
 import { model } from './model.js';
 import { setports } from './setports.js';
 import { queue } from './queue.js';
+import { allow } from './allow.js';
 
 export interface Command {
   data: SlashCommandBuilder;
@@ -26,4 +27,5 @@ commands.set(autowork.data.name, autowork);
 commands.set(model.data.name, model);
 commands.set(setports.data.name, setports as Command);
 commands.set(queue.data.name, queue);
+commands.set(allow.data.name, allow);